Fix Opentracing functional tests
