+++ repose-aggregator/tests/functional-tests/src/integrationTest/configs/features/services/opentracing/common/log4j2-test.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Configuration packages="org.apache.logging.log4j.flume.appender">
+    <Appenders>
+        <RollingFile name="RollingFile" fileName="${repose.log.name}"
+                     filePattern="${repose.log.pattern}" immediateFlush="true">
+            <PatternLayout pattern="%d %-4r [%t] %-5p %c - %m%n"/>
+            <Policies>
+                <SizeBasedTriggeringPolicy size="200 MB"/>
+            </Policies>
+            <DefaultRolloverStrategy max="2"/>
+        </RollingFile>
+    </Appenders>
+    <Loggers>
+        <Root level="debug">
+            <AppenderRef ref="RollingFile"/>
+        </Root>
+        <Logger name="com.sun.jersey" level="off"/>
+        <Logger name="net.sf.ehcache" level="error"/>
+        <Logger name="org.apache" level="warn"/>
+        <Logger name="org.eclipse.jetty" level="off"/>
+        <Logger name="org.openrepose" level="debug"/>
+        <Logger name="org.rackspace.deproxy" level="info"/>
+        <Logger name="org.springframework" level="warn"/>
+        <Logger name="intrafilter-logging" level="info"/>
+        <!--<Logger name="org.openrepose.core.services.serviceclient.akka.impl.AkkaServiceClientImpl" level="trace"/>-->
+
+        <!-- I need debug info from JMX! -->
+        <!-- useful: https://docs.oracle.com/javase/8/docs/technotes/guides/jmx/logging.html -->
+        <!-- also useful: http://logging.apache.org/log4j/2.x/log4j-jul/index.html -->
+        <!-- set these to FINEST to get log analysis of when mbeans were started and where -->
+        <Logger name="javax.management" level="debug"/>
+        <Logger name="javax.management.remote" level="debug"/>
+
+        <!-- because adding the JUL bridge, I need to silence a bunch of internal sun stuff -->
+        <Logger name="com.sun.xml.internal" level="warn"/>
+    </Loggers>
+</Configuration>
