--- repose-aggregator/tests/functional-tests/src/integrationTest/configs/features/services/opentracing/disabled/http/opentracing.cfg.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<opentracing xmlns="http://docs.openrepose.org/repose/opentracing/v1.0"
-             enabled="false"
-             name="repose"
-             tracer="jaeger"
-             collector-endpoint="http://localhost:${tracingPort}"
-/>
