+++ repose-aggregator/tests/functional-tests/src/integrationTest/configs/features/services/opentracing/enabledprobabilisticpoint9/http/open-tracing.cfg.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<open-tracing xmlns="http://docs.openrepose.org/repose/open-tracing-service/v1.0"
+             service-name="repose">
+    <jaeger>
+        <connection-http endpoint="http://localhost:${collectorTracingPort}" />
+        <sampling-probabilistic probability="0.9" />
+    </jaeger>
+</open-tracing>
