--- repose-aggregator/tests/functional-tests/src/integrationTest/configs/features/services/opentracing/enabledratelimiting/udp/opentracing.cfg.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<opentracing xmlns="http://docs.openrepose.org/repose/opentracing/v1.0"
-             name="repose"
-             tracer="jaeger"
-             sender-protocol="udp"
-             agent-port="${agentTracingPort}"
-             agent-host="localhost">
-    <jaeger-sampling-config sample-type="rate-limited">
-        <jaeger-sampling-rate-limiting max-traces-per-second="5.0" />
-    </jaeger-sampling-config>
-</opentracing>
