+++ repose-aggregator/tests/functional-tests/src/integrationTest/configs/features/services/opentracing/misconfigured/open-tracing.cfg.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<open-tracing xmlns="http://docs.openrepose.org/repose/open-tracing-service/v1.0"
+             service-name="repose">
+</open-tracing>
