--- repose-aggregator/tests/functional-tests/src/integrationTest/configs/features/services/opentracing/misconfigured/opentracing.cfg.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<opentracing xmlns="http://docs.openrepose.org/repose/opentracing/v1.0"
-             enabled="true"
-             name="repose"
-             tracer="unreal"
-/>
