--- repose-aggregator/tests/functional-tests/src/integrationTest/configs/features/services/opentracing/misconfiguredsampling/opentracing.cfg.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<opentracing xmlns="http://docs.openrepose.org/repose/opentracing/v1.0"
-             enabled="true"
-             name="repose"
-             tracer="jaeger">
-    <jaeger-sampling-config sample-type="rate-limited">
-        <jaeger-sampling-const value="1" />
-    </jaeger-sampling-config>
-</opentracing>
