--- repose-aggregator/tests/functional-tests/src/integrationTest/configs/features/services/opentracing/withkeystone/udp/opentracing.cfg.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<opentracing xmlns="http://docs.openrepose.org/repose/opentracing/v1.0"
-             name="repose"
-             tracer="jaeger"
-             sender-protocol="udp"
-             agent-port="${agentTracingPort}"
-             agent-host="localhost">
-</opentracing>
