+++ repose-aggregator/tests/functional-tests/src/integrationTest/groovy/features/services/opentracing/OpenTracingServiceHappyPathUdpTest.groovy
@@ -90,7 +90,7 @@ class OpenTracingServiceHappyPathUdpTest extends ReposeValveTest {
         when: "Request is sent through repose"
         def messageChain = deproxy.makeRequest(
             url: reposeEndpoint,
-            headers: ['uber-trace-id': trace_id ],
+            headers: [(TRACING_HEADER): trace_id ],
             method: method)
 
         then: "The request should have reached the origin service"
@@ -150,7 +150,7 @@ class OpenTracingServiceHappyPathUdpTest extends ReposeValveTest {
         when: "Request is sent through repose"
         def messageChain = deproxy.makeRequest(
             url: reposeEndpoint,
-            headers: ['uber-trace-id': trace_id ],
+            headers: [(TRACING_HEADER): trace_id ],
             method: method)
 
         then: "The request should have reached the origin service"
