+++ repose-aggregator/tests/functional-tests/src/integrationTest/groovy/features/services/opentracing/OpenTracingServiceNoConfigTest.groovy
@@ -80,7 +80,7 @@ class OpenTracingServiceNoConfigTest extends ReposeValveTest {
         when: "Request is sent through repose"
         def messageChain = deproxy.makeRequest(
             url: reposeEndpoint,
-            headers: ['uber-trace-id': trace_id ],
+            headers: [(TRACING_HEADER): trace_id ],
             method: method)
 
         then: "The request should have reached the origin service"
